No exception when trying to delete unknown service (just ERROR log) - Fix #131 (CF sometimes tries to delete services that failed to create, just to be sure)

+++ src/main/java/org/cloudfoundry/autosleep/ui/servicebroker/service/AutosleepServiceInstanceService.java
@@ -178,7 +178,11 @@ public class AutosleepServiceInstanceService implements ServiceInstanceService {
             DeleteServiceInstanceRequest request) throws ServiceBrokerException {
         final String spaceEnrollerConfigId = request.getServiceInstanceId();
         log.debug("deleteServiceInstance - {}", spaceEnrollerConfigId);
-        spaceEnrollerConfigRepository.delete(spaceEnrollerConfigId);
+        if (spaceEnrollerConfigRepository.findOne(spaceEnrollerConfigId) != null) {
+            spaceEnrollerConfigRepository.delete(spaceEnrollerConfigId);
+        } else {
+            log.error("Received delete on unknown id %s - This should not happen.", spaceEnrollerConfigId);
+        }
 
         //clean stored app linked to the service (already unbound)
         appRepository.findAll().forEach(
